+++ test/model.populate.test.js
@@ -9422,18 +9422,22 @@ describe('model: populate:', function() {
       const nested = Schema({}, { discriminatorKey: 'type' });
       const mainSchema = Schema({ items: [nested] });
 
+      const docs = [];
       mainSchema.path('items').discriminator('TestDiscriminator', Schema({
         childModel: { type: String },
         child: {
           type: mongoose.Schema.Types.ObjectId,
-          refPath: (doc, path) => path.replace('.child', '.childModel')
+          refPath: (doc, path) => {
+            docs.push(doc);
+            return path.replace('.child', '.childModel');
+          }
         }
       }));
       const Parent = db.model('Parent', mainSchema);
       const Child = db.model('Child', Schema({ name: String }));
 
       const child = yield Child.create({ name: 'test' });
-      yield Parent.create({
+      const parent = yield Parent.create({
         items: [{
           type: 'TestDiscriminator',
           childModel: 'Child',
@@ -9441,9 +9445,14 @@ describe('model: populate:', function() {
         }]
       });
 
+      assert.equal(docs.length, 0);
       const doc = yield Parent.findOne().populate('items.child').exec();
+
       assert.equal(doc.items[0].child.name, 'test');
       assert.ok(doc.items[0].populated('child'));
+
+      assert.equal(docs.length, 1);
+      assert.equal(docs[0]._id.toHexString(), parent.items[0]._id.toHexString());
     });
   });
 
@@ -10521,6 +10530,59 @@ describe('model: populate:', function() {
 
       assert.equal(populatedBooks.length, 1);
       assert.equal(populatedBooks[0].author.name, 'Author1');
+    });
+  });
+
+  it('calls subdocument ref functions with subdocument as context (gh-8469)', function() {
+    const ImageSchema = Schema({ imageName: String });
+    const Image = db.model('Image', ImageSchema);
+
+    const TextSchema = Schema({ textName: String });
+    const Text = db.model('Text', TextSchema);
+
+    const opts = { _id: false };
+    let contexts = [];
+    const ItemSchema = Schema({
+      data: {
+        type: 'ObjectId',
+        ref: function() {
+          contexts.push(this);
+          return this.objectType;
+        }
+      },
+      objectType: String
+    }, opts);
+
+    const ExampleSchema = Schema({ test: String, list: [ItemSchema] });
+    const Example = db.model('Test', ExampleSchema);
+
+    return co(function*() {
+      const text = yield Text.create({ textName: 'test' });
+      const image = yield Image.create({ imageName: 'test' });
+      yield Example.create({
+        list: [{ data: text._id, objectType: 'Text' }, { data: image._id, objectType: 'Image' }]
+      });
+
+      assert.equal(contexts.length, 0);
+      let res = yield Example.findOne().populate('list.data');
+
+      assert.equal(contexts.length, 2);
+      assert.equal(contexts[0].objectType, 'Text');
+      assert.equal(contexts[1].objectType, 'Image');
+
+      assert.equal(res.list[0].data.textName, 'test');
+      assert.equal(res.list[1].data.imageName, 'test');
+
+      contexts = [];
+      res = yield Example.findOne();
+      yield res.populate('list.data');
+
+      assert.equal(contexts.length, 2);
+      assert.equal(contexts[0].objectType, 'Text');
+      assert.equal(contexts[1].objectType, 'Image');
+
+      assert.equal(res.list[0].data.textName, 'test');
+      assert.equal(res.list[1].data.imageName, 'test');
     });
   });
 });